requirements.txt
fastapi
uvicorn
requests
python-dotenv
networkx
pymupdf
tqdm
chromadb
sentence-transformers